better handling of negotiated rates
negotiated rates are only included in the response if the NegotiatedRaesIndicator is set AND the Shipper Account/UserID is qualified to receive negotiated rates.

if negotiated_rates is set, but are not returned in the response, a notice is set warning that the account/userid is not qualified.

I believe the only way to have an account/userid qualified is to contact UPS support